Move plgin framework provider files to the new packages (#8160)
