+++ google/framework_provider_test.go
@@ -28,7 +28,7 @@ func GetFwTestProvider(t *testing.T) *frameworkTestProvider {
 
 	var diags diag.Diagnostics
 	p := NewFrameworkTestProvider(t.Name())
-	configureApiClient(context.Background(), &p.frameworkProvider, &diags)
+	configureApiClient(context.Background(), &p.FrameworkProvider, &diags)
 	if diags.HasError() {
 		log.Fatalf("%d errors when configuring test provider client: first is %s", diags.ErrorsCount(), diags.Errors()[0].Detail())
 	}
