+++ google/services/dns/data_source_dns_record_set.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package google
+package dns
 
 import (
 	"context"
@@ -13,6 +13,9 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
+	"github.com/hashicorp/terraform-provider-google/google/fwmodels"
+	"github.com/hashicorp/terraform-provider-google/google/fwresource"
+	"github.com/hashicorp/terraform-provider-google/google/fwtransport"
 )
 
 // Ensure the implementation satisfies the expected interfaces
@@ -90,22 +93,22 @@ func (d *GoogleDnsRecordSetDataSource) Configure(ctx context.Context, req dataso
 		return
 	}
 
-	p, ok := req.ProviderData.(*frameworkProvider)
+	p, ok := req.ProviderData.(*fwtransport.FrameworkProviderConfig)
 	if !ok {
 		resp.Diagnostics.AddError(
 			"Unexpected Data Source Configure Type",
-			fmt.Sprintf("Expected *frameworkProvider, got: %T. Please report this issue to the provider developers.", req.ProviderData),
+			fmt.Sprintf("Expected *fwtransport.FrameworkProviderConfig, got: %T. Please report this issue to the provider developers.", req.ProviderData),
 		)
 		return
 	}
 
-	d.client = p.NewDnsClient(p.userAgent, &resp.Diagnostics)
-	d.project = p.project
+	d.client = p.NewDnsClient(p.UserAgent, &resp.Diagnostics)
+	d.project = p.Project
 }
 
 func (d *GoogleDnsRecordSetDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
 	var data GoogleDnsRecordSetModel
-	var metaData *ProviderMetaModel
+	var metaData *fwmodels.ProviderMetaModel
 	var diags diag.Diagnostics
 
 	// Read Provider meta into the meta model
@@ -114,7 +117,7 @@ func (d *GoogleDnsRecordSetDataSource) Read(ctx context.Context, req datasource.
 		return
 	}
 
-	d.client.UserAgent = generateFrameworkUserAgentString(metaData, d.client.UserAgent)
+	d.client.UserAgent = fwtransport.GenerateFrameworkUserAgentString(metaData, d.client.UserAgent)
 
 	// Read Terraform configuration data into the model
 	resp.Diagnostics.Append(req.Config.Get(ctx, &data)...)
@@ -122,7 +125,7 @@ func (d *GoogleDnsRecordSetDataSource) Read(ctx context.Context, req datasource.
 		return
 	}
 
-	data.Project = getProjectFramework(data.Project, d.project, &resp.Diagnostics)
+	data.Project = fwresource.GetProjectFramework(data.Project, d.project, &resp.Diagnostics)
 	if resp.Diagnostics.HasError() {
 		return
 	}
@@ -130,7 +133,7 @@ func (d *GoogleDnsRecordSetDataSource) Read(ctx context.Context, req datasource.
 	data.Id = types.StringValue(fmt.Sprintf("projects/%s/managedZones/%s/rrsets/%s/%s", data.Project.ValueString(), data.ManagedZone.ValueString(), data.Name.ValueString(), data.Type.ValueString()))
 	clientResp, err := d.client.ResourceRecordSets.List(data.Project.ValueString(), data.ManagedZone.ValueString()).Name(data.Name.ValueString()).Type(data.Type.ValueString()).Do()
 	if err != nil {
-		handleDatasourceNotFoundError(ctx, err, &resp.State, fmt.Sprintf("dataSourceDnsRecordSet %q", data.Name.ValueString()), &resp.Diagnostics)
+		fwtransport.HandleDatasourceNotFoundError(ctx, err, &resp.State, fmt.Sprintf("dataSourceDnsRecordSet %q", data.Name.ValueString()), &resp.Diagnostics)
 		if resp.Diagnostics.HasError() {
 			return
 		}
