+++ google/services/firebase/data_source_google_firebase_web_app_config.go
@@ -1,3 +1,3 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package google
+package firebase
